Split out BulbId
