+++ lib/Types/BulbId.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include <stdint.h>
+#include <MiLightRemoteType.h>
+
+struct BulbId {
+  uint16_t deviceId;
+  uint8_t groupId;
+  MiLightRemoteType deviceType;
+
+  BulbId();
+  BulbId(const BulbId& other);
+  BulbId(const uint16_t deviceId, const uint8_t groupId, const MiLightRemoteType deviceType);
+  bool operator==(const BulbId& other);
+  void operator=(const BulbId& other);
+};
